a3/requirements_colab.txt
pandas
transformers
datasets
evaluate
torchmetrics
torchvision
tqdm
sentencepiece